--- a/.pre-commit-hooks.yaml
+++ b/.pre-commit-hooks.yaml
@@ -45,6 +45,7 @@
 - id: terraform_tflint
   name: Terraform validate with tflint
   description: Validates all Terraform configuration files with TFLint.
+  require_serial: true
   entry: terraform_tflint.sh
   language: script
   files: (\.tf|\.tfvars)$