--- a/terraform_tflint.sh
+++ b/terraform_tflint.sh
@@ -61,10 +61,7 @@ tflint_() {
 
   for path_uniq in $(echo "${paths[*]}" | tr ' ' '\n' | sort -u); do
     path_uniq="${path_uniq//__REPLACED__SPACE__/ }"
-
-    pushd "$path_uniq" > /dev/null
-    tflint "${ARGS[@]}"
-    popd > /dev/null
+    tflint "${ARGS[@]}" $path_uniq
   done
 }