This PR fixes a couple of problems I've experienced when using the `terraform_tflint` hook:
- Same directory tested multiple times due to the method that  `pre-commit` uses to split input files in batches: this is fixed with `require_serial: true`
- Incomplete path to the file with the error reported by `tflint`: this is fixed by removing the `pushd`/`popd` commands.

Note: Since these are a couple of small issues, I haven't opened any issue against the project. If this is required to get the PR merged, just let me know, and I'll create the issue.