@MaxymVlasov 2021-09-30
<details><summary>Tests</summary>

With serial:

```bash
for i in {1..200}; do { echo -e "\n\nTest run $i times\n\n";  time pre-commit try-repo -a /mnt/c/Users/vm/code/open-source/pre-commit-terraform  terraform_tflint ; } 2>> terraform_tflint_with_require_serial_true; done

cat terraform_tflint_with_require_serial_true | awk '{ print $1; }' | datamash max 1 min 1 mean 1 median 1
```

Without serial:

```bash
terraform_tflint_without_serial 200 runs
for i in {1..200}; do { echo -e "\n\nTest run $i times\n\n";  time pre-commit try-repo -a /mnt/c/Users/vm/code/open-source/pre-commit-terraform  terraform_tflint ; } 2>> terraform_tflint_with_require_serial_true; done

cat terraform_tflint_without_serial | awk '{ print $1; }' | datamash max 1 min 1 mean 1 median 1
```

</details>

200 runs **with** serial:

| time command   | max    | min    | mean     | median |
| -------------- | ------ | ------ | -------- | ------ |
| users seconds  | 20.56  | 12.43  | 14.5878  | 14.14  |
| system seconds | 29.18  | 20.29  | 22.69225 | 22.145 |
| CPU %          | 177    | 158    | 167.81   | 168    |
| Total time     | 28.927 | 19.212 | 22.15246 | 21.562 |


200 runs **without** serial:

| time command   | max    | min    | mean      | median  |
| -------------- | ------ | ------ | --------- | ------- |
| users seconds  | 130.3  | 101.46 | 113.12495 | 112.63  |
| system seconds | 135.47 | 102.29 | 115.0939  | 114.595 |
| CPU %          | 739    | 700    | 730.17    | 731     |
| Total time     | 36.779 | 28.237 | 31.236725 | 31.0375 |

## Conclusion

Adding `require_serial: true` _at least_ does not affect performance on a big codebase. 

FYI, @antonbabenko